Added preconfigured rust support
I intend on using this primarily for rust development. as such if i am going to have to manually tell the lsp server to talk to the future autocompletion plugin i am only going to use neovim for languages that i am commited to in the long run, as doing extra config work for something i will only mess around with for a week or two at most is just not worth it. at that point i might as well use vs-code.